requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
redis==5.0.0
python-jose==3.3.0
python-dotenv==1.0.0
requests==2.31.0
prometheus-client==0.19.0
pyyaml==6.0.1
python-multipart
